# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Code of Reality Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [users]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Auth token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /assessments:
    get:
      operationId: listAssessments
      tags: [assessments]
      summary: List user assessments
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [DIMENSIONAL_SCAN, COHERENCE_AUDIT, IDENTITY_MAP, PHASE_TRANSITION]
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Assessment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assessment"
    post:
      operationId: submitAssessment
      tags: [assessments]
      summary: Submit a dimensional assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAssessmentBody"
      responses:
        "201":
          description: Assessment result with reality vector
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentResult"
...
  /assessments/{id}:
    get:
      operationId: getAssessment
      tags: [assessments]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Assessment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentResult"

  # ── REALITY VECTORS ────────────────────────────────────────────────────────
...
  /chapters:
    get:
      operationId: listChapters
      tags: [chapters]
      summary: List chapters (phase-gated by user's current phase)
      parameters:
        - in: query
          name: phase
          schema:
            type: integer
      responses:
        "200":
          description: Chapter list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chapter"
  /chapters/{id}:
    get:
      operationId: getChapter
      tags: [chapters]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chapter content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
  /chapters/{id}/complete:
    post:
      operationId: markChapterComplete
      tags: [chapters]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteChapterBody"
      responses:
        "200":
          description: Progress updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterProgress"
...
  /analytics/dimensions:
    get:
      operationId: getDimensionalTimeline
      tags: [analytics]
      summary: Get dimensional score timeline
      parameters:
        - in: query
          name: days
          schema:
            type: integer
            default: 90
      responses:
        "200":
          description: Dimensional timeline data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DimensionalTimeline"
...
components:
...
  schemas:
...
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        displayName:
          type: string
      required: [email, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [token, user]

    # ── USER ────────────────────────────────────────────────────────────────
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        currentPhase:
          type: integer
        onboardingDone:
          type: boolean
        subscriptionTier:
          type: string
          enum: [FREE, EXPLORER, ARCHITECT, CERTIFIED]
        streakDays:
          type: integer
        totalSessions:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, email, currentPhase, onboardingDone, subscriptionTier, streakDays, totalSessions, createdAt]
...
    DimensionScores:
      type: object
      description: Map of dimension index (0-7) to score (1-10)
      properties:
        spatial:
          type: number
        temporal:
          type: number
        causal:
          type: number
        intentional:
          type: number
        identity:
          type: number
        emotional:
          type: number
        mental:
          type: number
        transcendent:
          type: number
      required: [spatial, temporal, causal, intentional, identity, emotional, mental, transcendent]
    SubmitAssessmentBody:
      type: object
      properties:
        type:
          type: string
          enum: [DIMENSIONAL_SCAN, COHERENCE_AUDIT, IDENTITY_MAP, PHASE_TRANSITION]
        dimensionScores:
          $ref: "#/components/schemas/DimensionScores"
        responses:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: string
              score:
                type: number
              response:
                type: string
            required: [questionId, score]
        phase:
          type: integer
      required: [type, dimensionScores]
    Assessment:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [DIMENSIONAL_SCAN, COHERENCE_AUDIT, IDENTITY_MAP, PHASE_TRANSITION]
        dimensionScores:
          $ref: "#/components/schemas/DimensionScores"
        overallScore:
          type: number
        phase:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, type, dimensionScores, overallScore, phase, createdAt]
    AssessmentResult:
      type: object
      properties:
        assessment:
          $ref: "#/components/schemas/Assessment"
        realityVector:
          $ref: "#/components/schemas/RealityVector"
        patternLocks:
          type: array
          items:
            $ref: "#/components/schemas/PatternLock"
        insights:
          type: array
          items:
            type: string
      required: [assessment, realityVector, patternLocks, insights]

    # ── REALITY VECTOR ────────────────────────────────────────────────────────
    RealityVector:
      type: object
      properties:
        id:
          type: string
        components:
          type: array
          items:
            type: number
          minItems: 8
          maxItems: 8
        magnitude:
          type: number
        direction:
          type: array
          items:
            type: number
        trajectory:
          type: array
          items:
            type: number
        stability:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, components, magnitude, direction, trajectory, stability, createdAt]
...
    PatternLock:
      type: object
      properties:
        id:
          type: string
        dimension:
          type: integer
        dimensionName:
          type: string
        description:
          type: string
        severity:
          type: number
        status:
          type: string
          enum: [LOCKED, UNLOCKING, RESOLVED]
        timelineOrigin:
          type: string
          nullable: true
        resolvedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, dimension, dimensionName, description, severity, status, createdAt]
...
    Chapter:
      type: object
      properties:
        id:
          type: string
        phase:
          type: integer
        sequence:
          type: integer
        title:
          type: string
        openingParagraph:
          type: string
        summary:
          type: string
        quotables:
          type: array
          items:
            type: string
        linkedToolIds:
          type: array
          items:
            type: string
        isUnlocked:
          type: boolean
        progress:
          $ref: "#/components/schemas/ChapterProgress"
          nullable: true
      required: [id, phase, sequence, title, openingParagraph, summary, quotables, linkedToolIds, isUnlocked]
    ChapterProgress:
      type: object
      properties:
        chapterId:
          type: string
        isComplete:
          type: boolean
        assessmentScore:
          type: number
          nullable: true
        reflectionSubmitted:
          type: boolean
        completedAt:
          type: string
          format: date-time
          nullable: true
        lastReadAt:
          type: string
          format: date-time
          nullable: true
      required: [chapterId, isComplete, reflectionSubmitted]
    CompleteChapterBody:
      type: object
      properties:
        assessmentScore:
          type: number
        reflectionText:
          type: string
      required: [assessmentScore]

    # ── PHASE ────────────────────────────────────────────────────────────────
...
    DimensionalTimeline:
      type: object
      properties:
        dataPoints:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date-time
              scores:
                $ref: "#/components/schemas/DimensionScores"
              overallScore:
                type: number
            required: [date, scores, overallScore]
        dimensionLabels:
          type: array
          items:
            type: string
        phaseTransitions:
          type: array
          items:
            type: object
            properties:
              phase:
                type: integer
              date:
                type: string
                format: date-time
            required: [phase, date]
      required: [dataPoints, dimensionLabels, phaseTransitions]
...